Remove deprecated CCF stability calculation
Move functions to extra_functions in case they have later use.
