+++ supreme_spoon/run_DMS.py
@@ -105,10 +105,7 @@ if 2 in config['run_stages']:
                                 smoothed_wlc=config['smoothed_wlc'],
                                 save_results=config['save_results'],
                                 force_redo=config['force_redo'],
-                                calculate_stability_ccf=config['calculate_stability_ccf'],
-                                stability_params_ccf=config['stability_params_ccf'],
-                                nthreads=config['nthreads'],
-                                calculate_stability_pca=config['calculate_stability_pca'],
+                                calculate_stability=config['calculate_stability'],
                                 pca_components=config['pca_components'],
                                 output_tag=config['output_tag'],
                                 smoothing_scale=config['smoothing_scale'],
